docs/api: Improve Golang API documentation
Embeed the existing examples on the documentation
to they can be rendered on the website.

This also performs a bit of cleanup to the examples
to make the clearer and more consistent.
