+++ examples/gadgets/datasource/.gitignore
