+++ examples/gadgets/datasource/main.go
@@ -19,33 +19,42 @@ import (
 	"fmt"
 	"os"
 
-	orasoci "oras.land/oras-go/v2/content/oci"
-
+	"github.com/inspektor-gadget/inspektor-gadget/pkg/datasource"
 	gadgetcontext "github.com/inspektor-gadget/inspektor-gadget/pkg/gadget-context"
-	clioperator "github.com/inspektor-gadget/inspektor-gadget/pkg/operators/cli"
+	"github.com/inspektor-gadget/inspektor-gadget/pkg/operators"
 	_ "github.com/inspektor-gadget/inspektor-gadget/pkg/operators/ebpf"
 	ocihandler "github.com/inspektor-gadget/inspektor-gadget/pkg/operators/oci-handler"
+	"github.com/inspektor-gadget/inspektor-gadget/pkg/operators/simple"
 	"github.com/inspektor-gadget/inspektor-gadget/pkg/runtime/local"
 )
 
 func do() error {
-	ctx := context.Background()
+	const opPriority = 50000
+	myOperator := simple.New("myOperator",
+		simple.OnInit(func(gadgetCtx operators.GadgetContext) error {
+			for _, d := range gadgetCtx.GetDataSources() {
+				// Access some specific fields
+				pidF := d.GetField("proc.pid")
+				commF := d.GetField("proc.comm")
+				fnameF := d.GetField("fname")
 
-	// Create the OCI store from a tarball.
-	// To create this tarball run:
-	// $ sudo ig image pull trace_open:latest
-	// $ sudo ig image export trace_open:latest /tmp/trace_open.tar
-	ociStore, err := orasoci.NewFromTar(ctx, "trace_open.tar")
-	if err != nil {
-		return fmt.Errorf("getting oci store from bundle: %w", err)
-	}
+				d.Subscribe(func(source datasource.DataSource, data datasource.Data) error {
+					// Error handling is omitted for brevity
+					pid, _ := pidF.Uint32(data)
+					comm, _ := commF.String(data)
+					fname, _ := fnameF.String(data)
+					fmt.Printf("command %s (%d) opened %s\n", comm, pid, fname)
+					return nil
+				}, opPriority)
+			}
+			return nil
+		}),
+	)
 
 	gadgetCtx := gadgetcontext.New(
 		context.Background(),
-		// The name of the gadget to run is needed as a tarball can contain multiple images.
 		"ghcr.io/inspektor-gadget/gadget/trace_open:latest",
-		gadgetcontext.WithDataOperators(ocihandler.OciHandler, clioperator.CLIOperator),
-		gadgetcontext.WithOrasReadonlyTarget(ociStore),
+		gadgetcontext.WithDataOperators(ocihandler.OciHandler, myOperator),
 	)
 
 	runtime := local.New()
@@ -54,10 +63,7 @@ func do() error {
 	}
 	defer runtime.Close()
 
-	params := map[string]string{
-		"operator.cli.output": "columns",
-	}
-	if err := runtime.RunGadget(gadgetCtx, nil, params); err != nil {
+	if err := runtime.RunGadget(gadgetCtx, nil, nil); err != nil {
 		return fmt.Errorf("running gadget: %w", err)
 	}
