+++ examples/gadgets/from_file/.gitignore
