+++ examples/gadgets/from_memory/.gitignore
