+++ examples/gadgets/from_memory/trace_open.tar
