+++ examples/gadgets/operators/.gitignore
