--- examples/gadgets/operators/cli/.gitignore
@@ -1,1 +1,0 @@
-cli
