+++ examples/gadgets/other/mutate/main.go
