+++ examples/gadgets/other/otel_metrics/.gitignore
