+++ examples/gadgets/other/otel_metrics/README.md
