+++ examples/gadgets/other/otel_metrics/main.go
