+++ examples/gadgets/trace_dns/.gitignore
