+++ examples/gadgets/trace_dns/README.md
