+++ examples/gadgets/trace_dns/main.go
