+++ examples/gadgets/trace_open/.gitignore
